Refactor AI/ML modules into dedicated components
