+++ src/ai_ml/engine.py
@@ -1,0 +1,69 @@
+"""Core execution and configuration utilities for AI/ML workflows."""
+
+from __future__ import annotations
+
+import json
+import logging
+from pathlib import Path
+from typing import Any, Dict, Optional
+
+from .workflows import MLWorkflow
+
+logger = logging.getLogger(__name__)
+
+
+class AIEngine:
+    """Provide configuration loading and workflow execution logic."""
+
+    def __init__(self, config_path: Optional[Path] = None):
+        self.config_path = (
+            Path(config_path) if config_path else Path("config/ai_ml/ai_ml.json")
+        )
+        self.config: Dict[str, Any] = self.load_config()
+        self.setup_logging()
+
+    # ------------------------------------------------------------------
+    # Configuration
+    # ------------------------------------------------------------------
+    def load_config(self) -> Dict[str, Any]:
+        """Load configuration from disk."""
+        try:
+            if self.config_path.exists():
+                with open(self.config_path, "r") as f:
+                    return json.load(f)
+            logger.warning(f"Config file not found: {self.config_path}")
+        except Exception as e:  # pragma: no cover - best effort logging
+            logger.error(f"Error loading config: {e}")
+        return {}
+
+    def setup_logging(self) -> None:
+        """Initialize logging based on configuration settings."""
+        try:
+            log_level = (
+                self.config.get("ai_ml", {}).get("logging", {}).get("level", "INFO")
+            )
+            logging.basicConfig(level=getattr(logging, log_level.upper()))
+        except Exception as e:  # pragma: no cover - best effort logging
+            logger.error(f"Failed to setup logging: {e}")
+
+    # ------------------------------------------------------------------
+    # Workflow Execution
+    # ------------------------------------------------------------------
+    def execute_workflow(self, workflow: MLWorkflow) -> bool:
+        """Execute the steps within a workflow."""
+        try:
+            workflow.status = "running"
+            for step in workflow.steps:
+                if step["status"] == "pending":
+                    step["status"] = "running"
+                    # Placeholder for actual step execution logic
+                    step["status"] = "completed"
+            workflow.status = "completed"
+            return True
+        except Exception as e:  # pragma: no cover - best effort logging
+            workflow.status = "failed"
+            logger.error(f"Workflow execution failed: {workflow.name} - {e}")
+            return False
+
+
+__all__ = ["AIEngine"]
